Implement new metaAllocation scheme using heap-based memory rather than end of file, avoiding need to copy allocators on file extend and preserving commit state across extension.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/JOURNAL_HA_BRANCH@3154 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/rwstore/Allocator.java
@@ -36,8 +36,8 @@ public interface Allocator extends Comparable {
   public boolean addressInRange(int addr);
   public boolean free(int addr, int size);
   public int alloc(RWStore store, int size);
-  public long getDiskAddr();
-  public void setDiskAddr(long addr);
+  public int getDiskAddr();
+  public void setDiskAddr(int addr);
   public long getPhysicalAddress(int offset);
   public int getPhysicalSize(int offset);
   public byte[] write();
